Move all calypso checkout test helpers to own file
